[refactoring] will move FilmListTypes to const
