+++ src/utils/const.js
@@ -1,5 +1,24 @@
 const emotions = ['smile', 'sleeping', 'puke', 'angry'];
 
+const FilmListTypes = {
+  ALL_MOVIES: {
+    title: 'All movies. Upcoming',
+    isHidden: true,
+    isExtra: false,
+  },
+  TOP_MOVIES: {
+    title: 'Top rated',
+    isHidden: false,
+    isExtra: true,
+  },
+  COMMENTED_MOVIES: {
+    title: 'Most commented',
+    isHidden: false,
+    isExtra: true,
+  },
+};
+
+
 const FilmControlTypes = {
   WATCHLIST: 'watchlist',
   WATCHED: 'watched',
@@ -20,6 +39,7 @@ const SortTypes = {
 };
 
 export {
+  FilmListTypes,
   emotions,
   FilmControlTypes,
   NavigationTypes,
